validator refactoring
extracted node's code to be aligned with spec
