+++ protocol/v2/ssv/validator/validator.go
@@ -2,47 +2,15 @@ package validator
 
 import (
 	"context"
-	"github.com/bloxapp/ssv/ibft/storage"
-	"github.com/bloxapp/ssv/protocol/v2/qbft/instance"
-	"sync/atomic"
-	"time"
-
-	specp2p "github.com/bloxapp/ssv-spec/p2p"
 	specqbft "github.com/bloxapp/ssv-spec/qbft"
 	specssv "github.com/bloxapp/ssv-spec/ssv"
 	spectypes "github.com/bloxapp/ssv-spec/types"
-	"github.com/pkg/errors"
-	"go.uber.org/zap"
-
-	"github.com/bloxapp/ssv/protocol/v2/message"
+	"github.com/bloxapp/ssv/ibft/storage"
 	"github.com/bloxapp/ssv/protocol/v2/ssv/msgqueue"
 	"github.com/bloxapp/ssv/protocol/v2/ssv/runner"
 	"github.com/bloxapp/ssv/protocol/v2/types"
-)
-
-// Options represents options that should be passed to a new instance of Validator.
-type Options struct {
-	Logger      *zap.Logger
-	Network     specqbft.Network
-	Beacon      specssv.BeaconNode
-	Storage     *storage.QBFTStores
-	SSVShare    *types.SSVShare
-	Signer      spectypes.KeyManager
-	DutyRunners runner.DutyRunners
-	Mode        Mode
-	FullNode    bool
-}
-
-func (o *Options) defaults() {
-	if o.Logger == nil {
-		o.Logger = zap.L()
-	}
-}
-
-// set of states for the controller
-const (
-	NotStarted uint32 = iota
-	Started
+	"github.com/pkg/errors"
+	"go.uber.org/zap"
 )
 
 // Validator represents an SSV ETH consensus validator Share assigned, coordinates duty execution and more.
@@ -66,18 +34,9 @@ type Validator struct {
 
 	Q msgqueue.MsgQueue
 
-	mode int32
-
-	State uint32
+	mode  uint32
+	state uint32
 }
-
-// Mode defines a mode Validator operates in.
-type Mode int32
-
-const (
-	ModeRW Mode = iota
-	ModeR
-)
 
 // NewValidator creates a new instance of Validator.
 func NewValidator(pctx context.Context, options Options) *Validator {
@@ -102,87 +61,11 @@ func NewValidator(pctx context.Context, options Options) *Validator {
 		Share:       options.SSVShare,
 		Signer:      options.Signer,
 		Q:           q,
-		mode:        int32(options.Mode),
-		State:       NotStarted,
+		mode:        uint32(options.Mode),
+		state:       uint32(NotStarted),
 	}
 
 	return v
-}
-
-// Start starts a Validator.
-func (v *Validator) Start() error {
-	if atomic.CompareAndSwapUint32(&v.State, NotStarted, Started) {
-		n, ok := v.Network.(specp2p.Subscriber)
-		if !ok {
-			return nil
-		}
-		identifiers := v.DutyRunners.Identifiers()
-		for _, identifier := range identifiers {
-			if err := v.loadLastHeight(identifier); err != nil {
-				v.logger.Warn("could not load highest", zap.String("identifier", identifier.String()), zap.Error(err))
-			}
-			if err := n.Subscribe(identifier.GetPubKey()); err != nil {
-				return err
-			}
-			go v.StartQueueConsumer(identifier, v.ProcessMessage)
-			go v.sync(identifier)
-		}
-	}
-	return nil
-}
-
-func (v *Validator) sync(mid spectypes.MessageID) {
-	ctx, cancel := context.WithCancel(v.ctx)
-	defer cancel()
-
-	// TODO: config?
-	interval := time.Second
-	retries := 3
-
-	for ctx.Err() == nil {
-		err := v.Network.SyncHighestDecided(mid)
-		if err != nil {
-			v.logger.Debug("could not sync highest decided", zap.String("identifier", mid.String()))
-			retries--
-			if retries > 0 {
-				interval *= 2
-				time.Sleep(interval)
-				continue
-			}
-		}
-		return
-	}
-}
-
-// Stop stops a Validator.
-func (v *Validator) Stop() error {
-	v.cancel()
-	if atomic.LoadInt32(&v.mode) == int32(ModeR) {
-		return nil
-	}
-	// clear the msg q
-	v.Q.Clean(func(index msgqueue.Index) bool {
-		return true
-	})
-	return nil
-}
-
-// HandleMessage handles a spectypes.SSVMessage.
-func (v *Validator) HandleMessage(msg *spectypes.SSVMessage) {
-	if atomic.LoadInt32(&v.mode) == int32(ModeR) {
-		err := v.ProcessMessage(msg)
-		if err != nil {
-			v.logger.Warn("could not handle msg", zap.Error(err))
-		}
-		return
-	}
-	fields := []zap.Field{
-		zap.Int("queue_len", v.Q.Len()),
-		zap.String("msgType", message.MsgTypeToString(msg.MsgType)),
-		zap.String("msgID", msg.MsgID.String()),
-	}
-	v.logger.Debug("got message, add to queue", fields...)
-	v.Q.Add(msg)
 }
 
 // StartDuty starts a duty for the validator
@@ -237,28 +120,5 @@ func (v *Validator) validateMessage(runner runner.Runner, msg *spectypes.SSVMess
 		return errors.New("msg data is invalid")
 	}
 
-	return nil
-}
-
-func (v *Validator) loadLastHeight(identifier spectypes.MessageID) error {
-	storage := v.Storage.Get(identifier.GetRoleType())
-	if storage == nil {
-		return errors.New("storage not found")
-	}
-	highestState, err := storage.GetHighestInstance(identifier[:])
-	if err != nil {
-		return errors.Wrap(err, "failed to get heights instance state")
-	}
-	if highestState == nil {
-		return nil
-	}
-	r := v.DutyRunners.DutyRunnerForMsgID(identifier)
-	if r == nil {
-		return errors.New("runner is nil")
-	}
-	instance := instance.NewInstanceFromState(r.GetBaseRunner().QBFTController.GetConfig(), highestState)
-	r.GetBaseRunner().QBFTController.Height = instance.GetHeight()
-	r.GetBaseRunner().QBFTController.StoredInstances.AddNewInstance(instance)
-	v.logger.Info("highest instance loaded", zap.String("role", identifier.GetRoleType().String()), zap.Int64("h", int64(instance.GetHeight())))
 	return nil
 }
